requirements.txt
torch
torchvision
scikit-learn
matplotlib
numpy
Pillow
tqdm
